requirements.txt
asgiref==3.9.1
attrs==25.3.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.13.5
blinker==1.7.0
Brlapi==0.8.5
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.6
cloud-init==25.1.4
cloudinary==1.44.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==46.0.3
cupshelpers==1.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
docopt==0.6.2
docutils==0.20.1
drf-yasg==1.21.11
ebaysdk==2.1.5
et_xmlfile==2.0.0
freezegun==1.2.1
geoip2==2.9.0
gyp==0.1
httplib2==0.20.4
idna==2.10
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libsass==0.23.0
louis==3.29.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
maxminddb==2.8.2
mdurl==0.1.2
mongoengine==0.29.1
netaddr==0.8.0
netifaces==0.11.0
num2words==0.5.10
numpy==2.3.2
oauthlib==3.2.2
ofxparse==0.21
olefile==0.46
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.4.0
polib==1.1.1
psutil==5.9.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.25.1
pycparser==2.23
pycups==2.0.1
pydot==1.4.2
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pymongo==4.13.2
pyparsing==3.2.4
PyPDF2==2.12.1
pypng==0.20220715.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-stdnum==1.19
pytz==2025.2
pyusb==1.2.1
pyxdg==0.28
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
reportlab==4.4.7
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rjsmin==1.2.0
setuptools==68.1.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
systemd-python==235
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==1.26.5
vobject==0.9.6.1
wadllib==1.3.6
Werkzeug==2.2.2
wheel==0.42.0
whitenoise==6.11.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.2.1
